feat: enable defining env vars via `webiny.project.ts`
